Revamp
Basically to achieve a common theme and to make a working burger menu i had to remove everythin that i had done, I have a working burger menu, navigation doesnt work, need to add pages. In need of a revamp of the main page. Main theme and burger menu added